Feature/Implement Regex validation layer
* Implemented Regex validation layer
* Put in orded filenames with context
* Fixed bug with RegexEmailPattern
* Refactored validatorResult methods
* Updated documentation
