+++ configuration.go
@@ -40,7 +40,7 @@ const (
 	DomainCharsSize               = `\A.{4,255}\z`
 	EmailCharsSize                = `\A.{6,255}\z`
 	RegexDomainPattern            = `(?i)[\p{L}0-9]+([\-.]{1}[\p{L}0-9]+)*\.\p{L}{2,63}`
-	RegexEmailPattern             = `(\A([\p{L}0-9]+[\w|\-.+]*)@(` + RegexDomainPattern + `)\z)`
+	RegexEmailPattern             = `(\A([\p{L}0-9]+[\W\w]*)@(` + RegexDomainPattern + `)\z)`
 	RegexDomainFromEmail          = `\A.+@(.+)\z`
 	RegexSMTPErrorBodyPattern     = `(?i).*550{1}.*(user|account|customer|mailbox).*`
 	RegexPortNumber               = `(6553[0-5]|655[0-2]\d|65[0-4](\d){2}|6[0-4](\d){3}|[1-5](\d){4}|[1-9](\d){0,3})`
